WIP: working through more tests
